Base and child classes improved
Dragon.cs's state is before Soma included his code.

Learned that the timing when to call the base.Start() is important or else, the player (and other game elements) will not update properly when it meets another enemy.

Setting the start to protected/public/private also affects the code somehow.