Main problems with dragon prefabs identified, player and dragon data storage implemented
Identified that prefabs' parameters/attributes changed when dragged into a scene.

System for storing both player and dragon data are fully implemented.

Player goes to its position before transitioning to the attack scene when a fight ends/if the player chooses to leave a fight and scene shifts to the basic scene again.