Overall serialization improvement
Fight system is already functional. Working on fusing the dragons and the player - the animation, and the logic.

With the current setup - without serialization incorporated in the inventory system, that is - continuing the fight system is quite difficult. So, I shifted to improving the inventory system instead.

Currently in the process of improving the stones inventory and the dragon inventory.