Dynamic animation system implementation, UI fix
Finally was able to fully implement the animation system I was envisioning for the enemy upon entry to the attack scene. animReady bool was not setting though, so I changed it to a trigger that is initially set to on. Quite strange though! Still could not find a reason as to why bool cannot be set.

The UI is now looking better than before. SubPanel is not vanising upon mouse exit and is actually updating to the actual stats of the dragons owned by the player! Explored the yield return null command for the first time and is looking good for what I want to achieve.